module: Move ip4/base-config headers to module
Moved header files to the module directory to separate an shared interface
used by sof and native loadable modules.
